refactor(sprint11): split logic/mod.rs impl Logic block
- email_impl.rs (231) : get_emails*, fetch_email, store_email_flag, move/read/starred/delete/archive
- mailbox_impl.rs (451) : select/search/expunge/copy/store, mailbox CRUD, subscribe
- calendar_impl.rs (194) : calendar_event CRUD
- mod.rs (1622→770) : setup + user + auth + impl LogicTrait
