+++ src/logic/mod.rs
@@ -388,862 +388,12 @@ impl Logic {
             })
         }
     }
-
-    pub async fn get_emails(&self, username: &str, mailbox: &str) -> Result<Vec<Email>> {
-        self.get_emails_page(username, mailbox, 200, 0).await
-    }
-
-    /// Sorted newest-first page; omits heavy headers projection at the driver level when possible.
-    pub async fn get_emails_page(
-        &self,
-        username: &str,
-        mailbox: &str,
-        limit: i64,
-        skip: u64,
-    ) -> Result<Vec<Email>> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Email>("emails");
-            let filter = doc! { "user_id": username, "mailbox": mailbox };
-            let cursor = collection
-                .find(filter)
-                .sort(doc! { "internal_date": -1 })
-                .skip(skip)
-                .limit(limit.max(1).min(200))
-                .await?;
-            cursor.try_collect().await
-        }
-        #[cfg(test)]
-        {
-            let _ = (limit, skip);
-            let emails = vec![Email::new(
-                "testemail",
-                "from@test.com",
-                "to@test.com",
-                "Test Subject",
-                "Test Body",
-            )];
-            Ok(emails)
-        }
-    }
-
-    pub async fn fetch_email(&self, username: &str, email_id: &str) -> Result<Option<Email>> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Email>("emails");
-            let filter = doc! { "user_id": username, "id": email_id };
-            collection.find_one(filter).await
-        }
-        #[cfg(test)]
-        {
-            //For test, we need to return an email
-            let email = Email::new(
-                "testemail",
-                "from@test.com",
-                "to@test.com",
-                "Test Subject",
-                "Test Body",
-            );
-            Ok(Some(email))
-        }
-    }
-
-    pub async fn store_email_flag(&self, username: &str, email_id: &str, flag: &str) -> Result<()> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Email>("emails");
-            let filter = doc! { "user_id": username, "id": email_id };
-            let update = doc! { "$addToSet": { "flags": flag } };
-            collection.update_one(filter, update).await?;
-            Ok(())
-        }
-        #[cfg(test)]
-        {
-            //For test, we need to return an empty result
-            Ok(())
-        }
-    }
-
-    pub async fn move_email_to_mailbox(
-        &self,
-        username: &str,
-        email_id: &str,
-        mailbox: &str,
-    ) -> Result<bool> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Email>("emails");
-            let filter = doc! { "user_id": username, "id": email_id };
-            let update = doc! { "$set": { "mailbox": mailbox.to_ascii_lowercase() } };
-            let res = collection.update_one(filter, update).await?;
-            Ok(res.matched_count > 0)
-        }
-        #[cfg(test)]
-        {
-            Ok(true)
-        }
-    }
-
-    pub async fn set_email_read(
-        &self,
-        username: &str,
-        email_id: &str,
-        is_read: bool,
-    ) -> Result<bool> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Email>("emails");
-            let filter = doc! { "user_id": username, "id": email_id };
-            let update = if is_read {
-                doc! { "$addToSet": { "flags": "\\Seen" } }
-            } else {
-                doc! { "$pull": { "flags": { "$in": ["\\Seen", "Seen", "seen"] } } }
-            };
-            let res = collection.update_one(filter, update).await?;
-            Ok(res.matched_count > 0)
-        }
-        #[cfg(test)]
-        {
-            Ok(true)
-        }
-    }
-
-    pub async fn set_email_starred(
-        &self,
-        username: &str,
-        email_id: &str,
-        is_starred: bool,
-    ) -> Result<bool> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Email>("emails");
-            let filter = doc! { "user_id": username, "id": email_id };
-            let update = if is_starred {
-                doc! { "$addToSet": { "flags": "\\Flagged" } }
-            } else {
-                doc! { "$pull": { "flags": { "$in": ["\\Flagged", "Flagged", "flagged", "starred"] } } }
-            };
-            let res = collection.update_one(filter, update).await?;
-            Ok(res.matched_count > 0)
-        }
-        #[cfg(test)]
-        {
-            Ok(true)
-        }
-    }
-
-    pub async fn delete_email(&self, username: &str, email_id: &str) -> Result<()> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Email>("emails");
-            let filter = doc! { "user_id": username, "id": email_id };
-            collection.delete_one(filter).await?;
-            Ok(())
-        }
-        #[cfg(test)]
-        {
-            self.client.delete_email(email_id).await
-        }
-    }
-
-    pub async fn archive_email(&self, username: &str, email_id: &str) -> Result<()> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Email>("emails");
-            let archive_collection = self
-                .client
-                .database(&database_name)
-                .collection::<Email>("archive");
-
-            let filter = doc! { "user_id": username, "id": email_id };
-            if let Some(document) = collection.find_one(filter.clone()).await? {
-                archive_collection.insert_one(document).await?;
-                collection.delete_one(filter).await?;
-                Ok(())
-            } else {
-                Err(Error::from(std::io::Error::new(
-                    std::io::ErrorKind::NotFound,
-                    "Email not found",
-                )))
-            }
-        }
-        #[cfg(test)]
-        {
-            self.client.archive_email(email_id).await
-        }
-    }
-
-    pub async fn select_mailbox(&self, username: &str, mailbox: &str) -> Result<Mailbox> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Mailbox>("mailboxes");
-
-            let filter = doc! { "user_id": username, "name": mailbox };
-            if let Some(mailbox) = collection.find_one(filter).await? {
-                Ok(mailbox)
-            } else {
-                Err(Error::from(std::io::Error::new(
-                    std::io::ErrorKind::NotFound,
-                    "Mailbox not found",
-                )))
-            }
-        }
-        #[cfg(test)]
-        {
-            Ok(Mailbox {
-                name: mailbox.to_string(),
-                flags: vec![],
-                exists: 10,
-                recent: 2,
-                unseen: 5,
-                permanent_flags: vec![],
-                uid_validity: 1,
-                uid_next: 11,
-                user_id: username.to_string(),
-            })
-        }
-    }
-
-    pub async fn search_messages(&self, username: &str, criteria: &str) -> Result<Vec<u32>> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Email>("emails");
-
-            let filter = match criteria {
-                "ALL" => doc! { "user_id": username },
-                "UNSEEN" => doc! { "user_id": username, "flags": { "$nin": ["\\Seen"] } },
-                "SEEN" => doc! { "user_id": username, "flags": "\\Seen" },
-                _ => doc! { "user_id": username },
-            };
-
-            let mut cursor = collection.find(filter).await?;
-            let mut sequence_numbers = Vec::new();
-            while let Some(email) = cursor.try_next().await? {
-                sequence_numbers.push(email.sequence_number);
-            }
-            Ok(sequence_numbers)
-        }
-        #[cfg(test)]
-        {
-            //For test, we need to return a vector of sequence numbers
-            let sequence_numbers = vec![1, 2, 3];
-            Ok(sequence_numbers)
-        }
-    }
-
-    pub async fn expunge_mailbox(&self, username: &str) -> Result<Vec<u32>> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Email>("emails");
-
-            let filter = doc! { "user_id": username, "flags": "\\Deleted" };
-            let mut cursor = collection.find(filter.clone()).await?;
-            let mut deleted_sequence_numbers = Vec::new();
-
-            while let Some(email) = cursor.try_next().await? {
-                deleted_sequence_numbers.push(email.sequence_number);
-            }
-
-            collection.delete_many(filter).await?;
-            Ok(deleted_sequence_numbers)
-        }
-        #[cfg(test)]
-        {
-            self.client.expunge_mailbox().await
-        }
-    }
-
-    pub async fn copy_messages(
-        &self,
-        username: &str,
-        sequence_set: &str,
-        _target_mailbox: &str,
-    ) -> Result<()> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Email>("emails");
-
-            let filter = doc! { "user_id": username, "sequence_number": sequence_set.parse::<u32>().unwrap_or(0) };
-            if let Some(mut email) = collection.find_one(filter).await? {
-                email.id = format!("{}_{}", email.id, Utc::now().timestamp());
-                collection.insert_one(email).await?;
-            }
-            Ok(())
-        }
-        #[cfg(test)]
-        {
-            self.client
-                .copy_messages(sequence_set, _target_mailbox)
-                .await
-        }
-    }
-
-    pub async fn store_flags(
-        &self,
-        username: &str,
-        sequence_set: &str,
-        flags: Vec<String>,
-        mode: &str,
-    ) -> Result<()> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Email>("emails");
-
-            let filter = doc! { "user_id": username, "sequence_number": sequence_set.parse::<u32>().unwrap_or(0) };
-            let update = match mode {
-                "+" => doc! { "$addToSet": { "flags": { "$each": flags } } },
-                "-" => doc! { "$pullAll": { "flags": flags } },
-                _ => doc! { "$set": { "flags": flags } },
-            };
-
-            collection.update_one(filter, update).await?;
-            Ok(())
-        }
-        #[cfg(test)]
-        {
-            self.client.store_flags(sequence_set, flags, mode).await
-        }
-    }
-
-    pub async fn check_mailbox(&self) -> Result<()> {
-        Ok(())
-    }
-
-    pub async fn close_mailbox(&self, username: &str) -> Result<()> {
-        self.expunge_mailbox(username).await?;
-        Ok(())
-    }
-
-    pub async fn noop(&self) -> Result<()> {
-        Ok(())
-    }
-
-    pub async fn get_mailbox_status(&self, username: &str, mailbox: &str) -> Result<Mailbox> {
-        self.select_mailbox(username, mailbox).await
-    }
-
-    pub async fn create_mailbox(&self, username: &str, mailbox: &str) -> Result<()> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Mailbox>("mailboxes");
-
-            // Vérifiez si la boîte aux lettres existe déjà
-            let mailbox_filter = doc! { "name": mailbox, "user_id": username };
-            if collection.find_one(mailbox_filter.clone()).await?.is_none() {
-                println!("Creating mailbox: {}", mailbox);
-                let new_mailbox = Mailbox {
-                    name: mailbox.to_string(),
-                    flags: vec![],
-                    exists: 0,
-                    recent: 0,
-                    unseen: 0,
-                    permanent_flags: vec![String::from("\\*")],
-                    uid_validity: 1,
-                    uid_next: 1,
-                    user_id: username.to_string(),
-                };
-                collection.insert_one(new_mailbox).await?;
-                println!("Mailbox created: {}", mailbox);
-            } else {
-                println!("Mailbox already exists: {}", mailbox);
-            }
-            Ok(())
-        }
-        #[cfg(test)]
-        {
-            // Pour les tests, nous devons retourner un résultat vide
-            Ok(())
-        }
-    }
-
-    pub async fn delete_mailbox(&self, username: &str, mailbox: &str) -> Result<()> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Mailbox>("mailboxes");
-
-            let filter = doc! { "user_id": username, "name": mailbox };
-            collection.delete_one(filter).await?;
-
-            Ok(())
-        }
-        #[cfg(test)]
-        {
-            //For test, we need to return an empty result
-            Ok(())
-        }
-    }
-
-    pub async fn rename_mailbox(
-        &self,
-        username: &str,
-        old_name: &str,
-        new_name: &str,
-    ) -> Result<()> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Mailbox>("mailboxes");
-
-            let filter = doc! { "user_id": username, "name": old_name };
-            let update = doc! { "$set": { "name": new_name } };
-            collection.update_one(filter, update).await?;
-            Ok(())
-        }
-        #[cfg(test)]
-        {
-            //For test, we need to return an empty result
-            Ok(())
-        }
-    }
-
-    pub async fn subscribe_mailbox(&self, username: &str, mailbox: &str) -> Result<()> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<User>("subscriptions");
-
-            let filter = doc! { "user_id": username, "mailbox": mailbox };
-            let update = doc! { "$set": { "subscribed": true } };
-            collection.update_one(filter, update).await?;
-            Ok(())
-        }
-        #[cfg(test)]
-        {
-            //For test, we need to return an empty result
-            Ok(())
-        }
-    }
-
-    pub async fn unsubscribe_mailbox(&self, username: &str, mailbox: &str) -> Result<()> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<User>("subscriptions");
-
-            let filter = doc! { "user_id": username, "mailbox": mailbox };
-            let update = doc! { "$set": { "subscribed": false } };
-            collection.update_one(filter, update).await?;
-            Ok(())
-        }
-        #[cfg(test)]
-        {
-            //For test, we need to return an empty result
-            Ok(())
-        }
-    }
-
-    pub async fn list_subscribed_mailboxes(
-        &self,
-        username: &str,
-        _reference: &str,
-        _pattern: &str,
-    ) -> Result<Vec<String>> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<User>("subscriptions");
-
-            let filter = doc! { "user_id": username, "subscribed": true };
-            let mut cursor = collection.find(filter).await?;
-            let mut mailboxes = Vec::new();
-
-            while let Some(subscription) = cursor.try_next().await? {
-                mailboxes.push(subscription.mailbox);
-            }
-
-            Ok(mailboxes)
-        }
-        #[cfg(test)]
-        {
-            self.client
-                .list_subscribed_mailboxes(username, _reference, _pattern)
-                .await
-        }
-    }
-
-    pub async fn get_mailbox_status_items(
-        &self,
-        username: &str,
-        mailbox: &str,
-        items: &str,
-    ) -> Result<String> {
-        #[cfg(not(test))]
-        {
-            let status = self.select_mailbox(username, mailbox).await?;
-            let mut response = Vec::new();
-
-            for item in items.split_whitespace() {
-                match item.trim_matches(|c| c == '(' || c == ')') {
-                    "MESSAGES" => response.push(format!("MESSAGES {}", status.exists)),
-                    "RECENT" => response.push(format!("RECENT {}", status.recent)),
-                    "UNSEEN" => response.push(format!("UNSEEN {}", status.unseen)),
-                    "UIDNEXT" => response.push(format!("UIDNEXT {}", status.uid_next)),
-                    "UIDVALIDITY" => response.push(format!("UIDVALIDITY {}", status.uid_validity)),
-                    _ => continue,
-                }
-            }
-
-            Ok(response.join(" "))
-        }
-        #[cfg(test)]
-        {
-            //For test, we need to return an empty result
-            Ok(String::new())
-        }
-    }
-
-    pub async fn store_email(&self, username: &str, mailbox: &str, email: &Email) -> Result<()> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<mongodb::bson::Document>("emails");
-
-            // Count existing emails to generate sequence_number and uid
-            let count = collection
-                .count_documents(doc! {"user_id": username, "mailbox": mailbox})
-                .await?;
-            let sequence_number = (count + 1) as u32;
-            let uid = (count + 1) as u32;
-
-            // Serialize the Email struct into a BSON document
-            let mut document = bson::to_document(email)?;
-            document.insert("user_id", username);
-            document.insert("mailbox", mailbox);
-            document.insert("sequence_number", sequence_number as i64);
-            document.insert("uid", uid as i64);
-            // Store internal_date as BSON datetime
-            document.insert(
-                "internal_date",
-                mongodb::bson::DateTime::from_millis(email.internal_date.timestamp_millis()),
-            );
-            // Insert the document into the collection
-            collection.insert_one(document).await?;
-            Ok(())
-        }
-        #[cfg(test)]
-        {
-            Ok(())
-        }
-    }
-
-    pub async fn list_mailboxes(
-        &self,
-        username: &str,
-        reference: &str,
-        mailbox: &str,
-    ) -> Result<Vec<String>> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<Mailbox>("mailboxes");
-
-            let filter = doc! {
-                "user_id": username,
-                "name": { "$regex": format!("^{}.*{}", reference, mailbox) }
-            };
-
-            let cursor = collection.find(filter).await?;
-            let mailboxes: Vec<String> = cursor.map_ok(|doc| doc.name).try_collect().await?;
-            Ok(mailboxes)
-        }
-        #[cfg(test)]
-        {
-            Ok(vec![
-                "inbox".to_string(),
-                "sent".to_string(),
-                "drafts".to_string(),
-            ])
-        }
-    }
-
-    // --- Calendar Event CRUD ---
-
-    pub async fn create_calendar_event(&self, event: &CalendarEvent) -> Result<()> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<CalendarEvent>("calendar_events");
-            collection.insert_one(event.clone()).await?;
-            Ok(())
-        }
-        #[cfg(test)]
-        {
-            Ok(())
-        }
-    }
-
-    pub async fn get_calendar_events(
-        &self,
-        username: &str,
-        start_after: Option<bson::DateTime>,
-        start_before: Option<bson::DateTime>,
-    ) -> Result<Vec<CalendarEvent>> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<CalendarEvent>("calendar_events");
-
-            let mut filter = doc! { "user_id": username };
-            if let Some(after) = start_after {
-                filter.insert("start", doc! { "$gte": after });
-            }
-            if let Some(before) = start_before {
-                if filter.contains_key("start") {
-                    if let Ok(start_doc) = filter.get_document_mut("start") {
-                        start_doc.insert("$lte", before);
-                    }
-                } else {
-                    filter.insert("start", doc! { "$lte": before });
-                }
-            }
-
-            let cursor = collection.find(filter).await?;
-            cursor.try_collect().await
-        }
-        #[cfg(test)]
-        {
-            Ok(vec![])
-        }
-    }
-
-    pub async fn get_calendar_event(
-        &self,
-        username: &str,
-        event_id: &str,
-    ) -> Result<Option<CalendarEvent>> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<CalendarEvent>("calendar_events");
-            let filter = doc! { "user_id": username, "id": event_id };
-            collection.find_one(filter).await
-        }
-        #[cfg(test)]
-        {
-            Ok(None)
-        }
-    }
-
-    pub async fn update_calendar_event(
-        &self,
-        username: &str,
-        event_id: &str,
-        update_doc: bson::Document,
-    ) -> Result<Option<CalendarEvent>> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<CalendarEvent>("calendar_events");
-            let filter = doc! { "user_id": username, "id": event_id };
-
-            let mut update = update_doc;
-            update.insert(
-                "updated_at",
-                bson::DateTime::from_millis(chrono::Utc::now().timestamp_millis()),
-            );
-
-            collection
-                .update_one(filter.clone(), doc! { "$set": update })
-                .await?;
-            collection.find_one(filter).await
-        }
-        #[cfg(test)]
-        {
-            Ok(None)
-        }
-    }
-
-    pub async fn delete_calendar_event(&self, username: &str, event_id: &str) -> Result<()> {
-        #[cfg(not(test))]
-        {
-            let database_name =
-                std::env::var("MONGODB_DATABASE").expect("MONGODB_DATABASE must be set");
-            let collection = self
-                .client
-                .database(&database_name)
-                .collection::<CalendarEvent>("calendar_events");
-            let filter = doc! { "user_id": username, "id": event_id };
-            collection.delete_one(filter).await?;
-            Ok(())
-        }
-        #[cfg(test)]
-        {
-            Ok(())
-        }
-    }
 }
 
-#[cfg(test)]
-use mockall::{automock, predicate::*};
+mod email_impl;
+mod mailbox_impl;
+mod calendar_impl;
 
-#[cfg_attr(test, automock)]
-#[async_trait::async_trait]
-pub trait DatabaseInterface: Send + Sync {
-    async fn insert_user(&self, user: User) -> Result<()>;
-    async fn find_user(&self, username: &str, password: &str) -> Result<Option<User>>;
-    async fn find_emails(&self, mailbox: &str) -> Result<Vec<Email>>;
-    async fn find_email(&self, email_id: &str) -> Result<Option<Email>>;
-    async fn update_email_flag(&self, email_id: &str, flag: &str) -> Result<()>;
-    async fn delete_email(&self, email_id: &str) -> Result<()>;
-    async fn archive_email(&self, email_id: &str) -> Result<()>;
-    async fn select_mailbox(&self, mailbox: &str) -> Result<Mailbox>;
-    async fn search_messages(&self, criteria: &str) -> Result<Vec<u32>>;
-    async fn expunge_mailbox(&self) -> Result<Vec<u32>>;
-    async fn copy_messages(&self, sequence_set: &str, target_mailbox: &str) -> Result<()>;
-    async fn store_flags(&self, sequence_set: &str, flags: Vec<String>, mode: &str) -> Result<()>;
-    async fn find_mailbox(&self, name: &str) -> Result<Option<Mailbox>>;
-    async fn update_mailbox(&self, mailbox: &str, update: Mailbox) -> Result<()>;
-    async fn create_mailbox(&self, mailbox: &str) -> Result<()>;
-    async fn delete_mailbox(&self, mailbox: &str) -> Result<()>;
-    async fn rename_mailbox(&self, old_name: &str, new_name: &str) -> Result<()>;
-    async fn subscribe_mailbox(&self, mailbox: &str) -> Result<()>;
-    async fn unsubscribe_mailbox(&self, mailbox: &str) -> Result<()>;
-    async fn list_subscribed_mailboxes(
-        &self,
-        username: &str,
-        reference: &str,
-        pattern: &str,
-    ) -> Result<Vec<String>>;
-    async fn get_mailbox_status_items(
-        &self,
-        username: &str,
-        mailbox: &str,
-        items: &str,
-    ) -> Result<String>;
-    async fn store_email(&self, username: &str, mailbox: &str, message: &str) -> Result<()>;
-    async fn get_mailbox_status(&self, username: &str, mailbox: &str) -> Result<Mailbox>;
-    async fn noop(&self) -> Result<()>;
-    async fn close_mailbox(&self) -> Result<()>;
-    async fn check_mailbox(&self) -> Result<()>;
-    async fn create_user(&self, username: &str, password: &str, mailbox: &str) -> Result<()>;
-    async fn authenticate_user(&self, username: &str, password: &str) -> Result<Option<User>>;
-    async fn list_mailboxes(
-        &self,
-        username: &str,
-        reference: &str,
-        mailbox: &str,
-    ) -> Result<Vec<String>>;
-}
-
-#[async_trait::async_trait]
-pub trait LogicTrait: Send + Sync {
-    async fn create_user(&self, username: &str, password: &str, mailbox: &str) -> Result<()>;
-}
 
 #[async_trait::async_trait]
 impl LogicTrait for Logic {
